Atualizando classe Temporaria de SellProduct para double

+++ src/main/java/me/davethecamper/cashshop/inventory/configs/temporary/TemporarySellProductMenu.java
@@ -16,11 +16,11 @@ public class TemporarySellProductMenu extends SellProductMenu {
 	private static final long serialVersionUID = -3241739116571529794L;
 	
 
-	public TemporarySellProductMenu(ItemMenuProperties item_properties, ProductConfig product, int updated_value, Consumer<TemporarySellProductMenu> callback) {
+	public TemporarySellProductMenu(ItemMenuProperties item_properties, ProductConfig product, double updated_value, Consumer<TemporarySellProductMenu> callback) {
 		this(item_properties, product, updated_value, 0, callback);
 	}
 
-	public TemporarySellProductMenu(ItemMenuProperties item_properties, ProductConfig product, int updated_value, long delay, Consumer<TemporarySellProductMenu> callback) {
+	public TemporarySellProductMenu(ItemMenuProperties item_properties, ProductConfig product, double updated_value, long delay, Consumer<TemporarySellProductMenu> callback) {
 		super("temporary_product", CashShop.getInstance().getMessagesConfig(), null, item_properties, product, updated_value, delay);
 		
 		this.callback = callback;
